+++ src/components/ui/text-effect.tsx
@@ -9,32 +9,7 @@ import {
   Variants,
 } from "motion/react";
 import React from "react";
-
-export type PresetType = "blur" | "fade-in-blur" | "scale" | "fade" | "slide";
-
-export type PerType = "word" | "char" | "line";
-
-export type TextEffectProps = {
-  children: string;
-  per?: PerType;
-  as?: keyof React.JSX.IntrinsicElements;
-  variants?: {
-    container?: Variants;
-    item?: Variants;
-  };
-  className?: string;
-  preset?: PresetType;
-  delay?: number;
-  speedReveal?: number;
-  speedSegment?: number;
-  trigger?: boolean;
-  onAnimationComplete?: () => void;
-  onAnimationStart?: () => void;
-  segmentWrapperClassName?: string;
-  containerTransition?: Transition;
-  segmentTransition?: Transition;
-  style?: React.CSSProperties;
-};
+import { PresetType, PerType, TextEffectProps } from "@/types/ui.types";
 
 const defaultStaggerTimes: Record<PerType, number> = {
   char: 0.03,
