+++ src/types/hooks.types.ts
@@ -1,0 +1,13 @@
+export interface UseResponsiveReturn {
+  isMobile: boolean;
+  isTablet: boolean;
+  isDesktop: boolean;
+}
+
+export interface PageSizes {
+  xs: number;
+  sm: number;
+  md: number;
+  lg: number;
+  xl: number;
+}
