+++ src/types/index.ts
@@ -1,0 +1,20 @@
+// Auth 관련 타입
+export * from "./auth.types";
+
+// 상품 관련 타입
+export * from "./products.types";
+
+// 댓글 관련 타입
+export * from "./comments.types";
+
+// 게시글 관련 타입
+export * from "./articles.types";
+
+// 컴포넌트 관련 타입
+export * from "./components.types";
+
+// UI 효과 관련 타입
+export * from "./ui.types";
+
+// 훅 관련 타입
+export * from "./hooks.types";
