+++ src/types/ui.types.ts
@@ -1,0 +1,35 @@
+import { ReactNode } from "react";
+import { Transition, Variants } from "motion/react";
+
+export type PresetType = "blur" | "fade-in-blur" | "scale" | "fade" | "slide";
+
+export type PerType = "word" | "char" | "line";
+
+export type TextEffectProps = {
+  children: string;
+  per?: PerType;
+  as?: keyof React.JSX.IntrinsicElements;
+  variants?: {
+    container?: Variants;
+    item?: Variants;
+  };
+  className?: string;
+  preset?: PresetType;
+  delay?: number;
+  speedReveal?: number;
+  speedSegment?: number;
+  trigger?: boolean;
+  onAnimationComplete?: () => void;
+  onAnimationStart?: () => void;
+  segmentWrapperClassName?: string;
+  containerTransition?: Transition;
+  segmentTransition?: Transition;
+  style?: React.CSSProperties;
+};
+
+export type BorderTrailProps = {
+  className?: string;
+  pathClassName?: string;
+  strokeWidth?: number;
+  children: ReactNode;
+};
