+++ src/utils/images/url.ts
@@ -1,4 +1,4 @@
-import { api } from "@/api/axios";
+import { api } from "@/services/axios";
 
 // 백엔드에서 전달된 이미지 경로를 완전한 URL로 변환
 export const getImageUrl = (
